A hack to support mutable hash-tables
The first table is mutated to become the merged dictionary. It is then returned
to the user. I think this is the natural way to merge mutable dictionaries. It
doesn't make sense to copy it and the mutate the new one because mutable
hash tables are generally not used that way.